Terms own their own value

+++ src/core/global.rs
@@ -2,7 +2,7 @@ use std::io::{BufWriter, Write};
 use std::io;
 
 pub type DocId = usize;
-pub type FieldId = u32;
+pub type FieldId = u8;
 
 #[derive(Clone,Debug,PartialEq,PartialOrd,Eq,Hash)]
 pub struct Field(pub FieldId);
